% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_nbc.R
\name{calculate_probability}
\alias{calculate_probability}
\title{Calculate P(B)}
\usage{
calculate_probability(x, feature, params_table)
}
\arguments{
\item{x}{numeric : relative abundance of the feature of interest}

\item{feature}{string : feature of interest with representation in the provided parameter table}

\item{pararms_table}{data frame / matrix : parameter table containing P(not detected) and shape parameters for beta distribution:}
}
\value{
numeric value
}
\description{
Calculate P(B)
}
