% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_fn.R
\name{clean}
\alias{clean}
\title{Uniformize raster layers.}
\usage{
clean(layers)
}
\arguments{
\item{layers}{Raster* object as defined by package raster.}
}
\value{
A Raster* object, same class as layers.
}
\description{
Crop raster layers to minimum size possible and uniformize NA values across layers.
}
\details{
Excludes all marginal rows and columns with only NA values and change values to NA if they are NA in any of the layers.
}
\examples{
data(gecko.layers)
raster::plot(clean(gecko.layers))
}
