% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier_fn.R
\name{confusion.matrix}
\alias{confusion.matrix}
\title{Create a confusion matrix}
\usage{
confusion.matrix(actual, predicted)
}
\arguments{
\item{actual}{dataframe. Original labels.}

\item{predicted}{dataframe. Predicted labels.}
}
\value{
data.frame. Predicted labels (rows) x Observed labels (cols).
}
\description{
Create a confusion matrix for any multiclass set of predicted vs observed labels
in a classification problem.
}
\examples{
x = c("FALSE", "TRUE", "FALSE", "TRUE", "TRUE")
y = c("TRUE", "TRUE", "TRUE", "TRUE", "TRUE")
confusion.matrix(x, y)
}
