% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concordance.R
\name{count_concordant_pairs}
\alias{count_concordant_pairs}
\title{Get proportion or count of pairwise concordant taxa per dyad}
\usage{
count_concordant_pairs(
  mat1_i,
  mat2_j,
  asv_table,
  feature_list,
  mat1,
  mat2,
  limit_of_detection,
  type = "proportion"
)
}
\arguments{
\item{mat1_i}{integer column i for matrix 1}

\item{mat2_j}{integer column j for matrix 2}

\item{asv_table}{features x samples: relative abundance}

\item{feature_list}{list of features to include}

\item{mat1}{class 1 binary sample concordance matrix (based on ROC thresholds from getCountMatrixByROCThreshold)}

\item{mat2}{class 2 binary sample concordance matrix (based on ROC thresholds from getCountMatrixByROCThreshold)}

\item{limit_of_detection}{relative abundance lower cutoff}

\item{type}{character option specifying proportion "proportion" or absolute "absolute" counts}
}
\description{
Get proportion or count of pairwise concordant taxa per dyad
}
