% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_fn.R
\name{create.east}
\alias{create.east}
\title{Create eastness layer.}
\usage{
create.east(layers)
}
\arguments{
\item{layers}{SpatRaster. A layer of elevation (a digital elevation model - DEM). 
As defined in package terra, see \code{\link[terra:rast]{terra::rast()}}.}
}
\value{
SpatRaster.
}
\description{
Create a layer depicting eastness based on an elevation layer.
}
\details{
Using elevation, aspect can be calculated. Yet, it is a circular variable (0 = 360) and has to be converted to northness and eastness to be useful for modelling.
}
\examples{
region = gecko.data("layers")
terra::plot(create.east(region[[3]]))
}
