% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_fn.R
\name{create.lat}
\alias{create.lat}
\title{Create latitude layer.}
\usage{
create.lat(layers)
}
\arguments{
\item{layers}{SpatRaster. As defined in package terra, see \code{\link[terra:rast]{terra::rast()}}.}
}
\value{
SpatRaster.
}
\description{
Create a layer depicting latitude based on any other.
}
\details{
Using latitude (and longitude) in models may help limiting the extrapolation of the predicted area much beyond known areas.
}
\examples{
region = gecko.data("layers")
terra::plot(create.lat(region[[1]]))
}
