% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_fn.R
\name{create.north}
\alias{create.north}
\title{Create northness layer.}
\usage{
create.north(dem)
}
\arguments{
\item{dem}{RasterLayer object of elevation (a digital elevation model - DEM) as defined by package raster.}
}
\value{
A RasterLayer object.
}
\description{
Create a layer depicting northness based on an elevation layer.
}
\details{
Using elevation, aspect can be calculated. Yet, it is a circular variable (0 = 360) and has to be converted to northness and eastness to be useful for modelling.
}
\examples{
data(gecko.layers)
raster::plot(create.north(gecko.layers[[3]]))
}
