% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concordance.R
\name{dyads_for_coverage}
\alias{dyads_for_coverage}
\title{Get Dyads needed to cover given list of features}
\usage{
dyads_for_coverage(
  ASVs_to_cover,
  perform_NBC_output,
  sampleConcordanceMatrix,
  allHumanASVs,
  limit_of_detection = 0,
  case_class_name = "asthmatic",
  ctrl_class_name = "healthy"
)
}
\arguments{
\item{ASVs_to_cover}{list of features for which you'd like to get coverage in a gnotobiotic exptt}

\item{perform_NBC_output}{perform_NBC_output}

\item{sampleConcordanceMatrix}{output of getCountMatrixByROCThreshold}

\item{allHumanASVs}{filtered ASV Table used for model training}

\item{limit_of_detection}{limit_of_detection (default=0)}

\item{case_class_name}{case_class_name (default = "asthmatic")}

\item{ctrl_class_name}{ctrl_class_name (default = "healthy")}
}
\value{
matrix of dyads best for full coverage of input feature list. Columns: Case (disease cohort) and Ctrl (healthy/control cohort)
}
\description{
Get Dyads needed to cover given list of features
}
