% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.R
\name{estimate_binomial_parameter}
\alias{estimate_binomial_parameter}
\title{Estimate Binomial Parameter}
\usage{
estimate_binomial_parameter(taxa, limit_of_detection)
}
\arguments{
\item{taxa}{vector of relative abundances per taxon}

\item{limit_of_detection}{numeric relative abundance value}
}
\value{
numeric weight value
}
\description{
This function estimates probability of non-detection per taxon:
Number of samples below the limit of detection divided by
the total number of samples.
A pseudocount of 1/2 is also included.
}
