\name{GECKO-package}
\alias{GECKO-package}
\alias{GECKO}
\docType{package}
\title{Genetic and Environmental Covariance Estimation by Composite-Likelihood Optimization
\packageTitle{GECKO}
}
\description{GECKO used the summary statistics from GWAS study to estimate genetic and environmental covariance between traits.
\packageDescription{GECKO}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{GECKO}
\packageIndices{GECKO}
Use GECKO_R function to estimate and test for genetic and environmental covariance. The inputs of the function including the summary statistics file for both traits, LD score, and the number of samples in each study. Annotation label is needed if annotation stratified genetic and environmental covariance are estimated 
}
\author{
\packageAuthor{GECKO}

Maintainer: \packageMaintainer{GECKO}
}
