% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_fn.R
\name{gecko.examples}
\alias{gecko.examples}
\title{Example data packaged with *gecko*}
\source{
This function is inspired by `palmerpanguins::path_to_file()`
which in turn is based on `readxl::readxl_example()`.
}
\usage{
gecko.examples(data = NULL)
}
\arguments{
\item{data}{Name of data in quotes. E.g.: `"gecko.records"`
If `NULL`, the example files will be listed.}
}
\description{
Load data included in the package. This includes *gecko.records*,
a matrix of longitude and latitude (two columns) occurrence records for
Hogna maderiana (Walckenaer, 1837); *gecko.range*, a SpatRaster object, as
defined by package terra, of the geographic range of Hogna maderiana
(Walckenaer, 1837); *gecko.layers*, a SpatRaster object with layers 
representing the average annual temperature, total annual precipitation,
altitude and landcover for Madeira Island
(Fick & Hijmans 2017, Tuanmu & Jetz 2014); and *worldborders* a vector of...
 World country borders.
}
\examples{
gecko.examples()
gecko.examples("gecko.range")
}
