% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\name{gecko.getDir}
\alias{gecko.getDir}
\title{Read GIS directory.}
\usage{
gecko.getDir()
}
\description{
Read directory where GIS files are stored.
}
\details{
Reads a txt file pointing to where the world GIS files are stored.
}
