% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_fn.R
\docType{data}
\name{gecko.range}
\alias{gecko.range}
\title{Geographic range for Hogna maderiana (Walckenaer, 1837).}
\format{
RasterLayer object as defined by package raster of range for Hogna maderiana (Walckenaer, 1837), a spider species from Madeira Island.
}
\usage{
data(gecko.range)
}
\description{
Geographic range for Hogna maderiana (Walckenaer, 1837).
}
\keyword{datasets}
