% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GECKO}
\alias{GECKO}
\title{A matrix-based RcppArmadillo implementation for computing x'Ay //'}
\usage{
GECKO(n1in, n2in, nsin, LDscorein, Zscorein, Annotation_labelin, Weightin,
  Fix_Vein, Test)
}
\arguments{
\item{x}{A size n vector}

\item{y}{A size m vector}

\item{A}{A size (n x m) matrix}
}
\value{
A scalar value evaluating x'Ay
}
\description{
A matrix-based RcppArmadillo implementation for computing x'Ay //'
}
