% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_fn.R
\docType{data}
\name{gecko.records}
\alias{gecko.records}
\title{Occurrence records for Hogna maderiana (Walckenaer, 1837).}
\format{
Matrix of longitude and latitude (two columns) of occurrence records for Hogna maderiana (Walckenaer, 1837), a spider species from Madeira Island.
}
\usage{
data(gecko.records)
}
\description{
Occurrence records for Hogna maderiana (Walckenaer, 1837).
}
\keyword{datasets}
