% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\name{gecko.setDir}
\alias{gecko.setDir}
\title{Setup GIS directory.}
\usage{
gecko.setDir(gisPath = NULL)
}
\arguments{
\item{gisPath}{Path to the directory where the gis files are stored.}
}
\description{
Setup directory where GIS files are stored.
}
\details{
Writes a txt file in the red directory allowing the package to always access the world GIS files directory.
}
