% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GECKO.R
\name{GECKO_R}
\alias{GECKO_R}
\title{GECKO}
\usage{
GECKO_R(Z1, Z2, n1in, n2in, nsin, LDscorein, Weightin, Fix_Vein, Test,
  Annotation_labelin = NULL)
}
\arguments{
\item{Z1}{summary statistics for trait 1. The summary statistic file should contain SNP ID, Major allele, and Z statistics with column name of SNP, A1 and Z}

\item{Z2}{summary statistics for trait 2. The summary statistic file should contain SNP ID, Major allele, and Z statistics with column name of SNP, A1 and Z}

\item{n1in}{Number of samples in study for trait 1}

\item{n2in}{Number of samples in study for trait 2}

\item{nsin}{Number of  overlapping samples of trait 1 and trait 2. If there is no overlapping sample, then the nsin is zero}

\item{LDscorein}{LD score for SNPs included in the study. The LD score should be pre-processed by LDSC}

\item{Weightin}{Whetehr to include weight for composite-likelihood. Suggested to be true to improve the estimation accuracy}

\item{Test}{Whethter to test for the genetic and environmental covariance.}

\item{Annotation_labelin}{Default is NULL, only need to be included if annotation-stratified genetic and environmental covariance are to be estimated}
}
\value{
List of model parameters including heritability, genetic covariance, environmental covariance, genetic correlation, and environmental correlation estimate and corresponding standard error and P value
}
\description{
Genetic and Environmental Covariance Estimation by Composite-Likelihood Optimization
}
\author{
Boran Gao, \email{borang@umich.edu}
}
