% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{get_concordance_matrices}
\alias{get_concordance_matrices}
\title{Get Pairwise Concordance}
\usage{
get_concordance_matrices(
  perform_NBC_output,
  sampleConcordanceMatrix,
  featureAUCsForConcordanceList,
  allHumanASVs,
  limit_of_detection,
  case_class_name = "asthmatic",
  ctrl_class_name = "healthy",
  PWCtype = "proportion"
)
}
\arguments{
\item{perform_NBC_output}{entire output from perform_NBC}

\item{sampleConcordanceMatrix}{binary matrix of all samples per taxon notifying sample concordance}

\item{featureAUCsForConcordanceList}{list of feature names: subset of ASVs we want to use to count pairwise concordance (top X, or AUC>0.6)}

\item{allHumanASVs}{ASV table for all samples}

\item{limit_of_detection}{relative abundance cutoff}

\item{PWCtype}{type of pairwise concordance count (proportion or absolute)}
}
\value{
list of matrices 1) proportion of pairwise concordant taxa per pair and 2) multidimensional binary matrix- one per case sample
}
\description{
Get Pairwise Concordance
}
