% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{get_modeled_features}
\alias{get_modeled_features}
\title{Get list of features included in model}
\usage{
get_modeled_features(model)
}
\arguments{
\item{model}{output from train_NBC}
}
\description{
Get list of features included in model
}
