% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{get_pairwise_concordant_features}
\alias{get_pairwise_concordant_features}
\title{Get full matrix of pairwise concordant features between 2 samples}
\usage{
get_pairwise_concordant_features(
  mat1_i,
  mat2_j,
  asv_table,
  feature_list,
  mat1,
  mat2,
  limit_of_detection
)
}
\arguments{
\item{mat1_i}{integer column i for matrix 1}

\item{mat2_j}{integer column j for matrix 2}

\item{asv_table}{features x samples: relative abundance}

\item{feature_list}{list of features to include}

\item{mat1}{class 1 binary sample concordance matrix (based on ROC thresholds from getCountMatrixByROCThreshold)}

\item{mat2}{class 2 binary sample concordance matrix (based on ROC thresholds from getCountMatrixByROCThreshold)}

\item{limit_of_detection}{relative abundance lower cutoff}
}
\description{
Get full matrix of pairwise concordant features between 2 samples
}
