% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{getBestThresholds}
\alias{getBestThresholds}
\title{Grab best thresholds per feature from ROC curve}
\usage{
getBestThresholds(
  sample_data,
  perform_NBC_output,
  train_NBC_output,
  featureAUCsRanked,
  idx1,
  idx2
)
}
\arguments{
\item{sample_data}{character matrix: column 1 = sample names; column 2 = class labels (e.g. "asthmatic" and "healthy")}

\item{perform_NBC_output}{entire output from perform_NBC}

\item{train_NBC_output}{entire output from train_NBC}

\item{featureAUCsRanked}{vector of AUCs in rank order}

\item{idx1}{starting rank for resulting list}

\item{idx2}{ending rank for resulting list}
}
\description{
Grab best thresholds per feature from ROC curve
}
