% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{getCountMatrixByROCThreshold}
\alias{getCountMatrixByROCThreshold}
\title{Get Concordance using ROC NB Score Thresholds}
\usage{
getCountMatrixByROCThreshold(
  perform_NBC_output,
  asvThresholds,
  ctrl_class_name = "healthy",
  case_class_name = "asthmatic"
)
}
\arguments{
\item{perform_NBC_output}{NBC output}

\item{asvThresholds}{best thresholds table with column "threshold", rownames feature names}

\item{ctrl_class_name}{default "healthy"}

\item{case_class_name}{default "asthmatic"}
}
\value{
binary matrix features vs samples 0=not concordant, 1=concordant, meaning healthy sample scored well for healthy for just that ASV value
}
\description{
Get Concordance using ROC NB Score Thresholds
}
