% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{getFeatureAUCs}
\alias{getFeatureAUCs}
\title{Get feature AUCs using pROC}
\usage{
getFeatureAUCs(train_NBC_output, perform_NBC_output, my_sample_data)
}
\arguments{
\item{train_NBC_output}{train_NBC output}

\item{perform_NBC_output}{perform_NBC output}

\item{my_sample_data}{character table col1=sample names, col2=class names}
}
\value{
table rownames=feature names, col "feature_aucs" containing numeric values from 0.5 to 1
}
\description{
Get feature AUCs using pROC
}
