% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.R
\name{getMapEstimates}
\alias{getMapEstimates}
\title{Maximum a posteriori estimation}
\usage{
getMapEstimates(taxa, sig_a = 30, sig_b = 100)
}
\arguments{
\item{taxa}{vector of relative abundances for a given taxon}

\item{sig_a}{Gaussian prior parameter for alpha (default=30)}

\item{sig_b}{Gaussian prior parameter for beta (default=100)}
}
\value{
fit beta parameters
}
\description{
Maximum a posteriori estimation
}
