% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{LogLoss}
\alias{LogLoss}
\title{Log Loss as defined on kaggle forum}
\usage{
LogLoss(act, pred)
}
\arguments{
\item{act}{actual value}

\item{pred}{predicted value}
}
\description{
Log Loss as defined on kaggle forum
}
