% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{make_confusion_matrix}
\alias{make_confusion_matrix}
\title{Make Confusion Matrix of Results}
\usage{
make_confusion_matrix(p1v2)
}
\arguments{
\item{p1v2}{second matrix in output of perform_NBC aka "sample_by_sample"}
}
\description{
Make Confusion Matrix of Results
}
