% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_gRNAs.R
\name{make_gRNAs}
\alias{make_gRNAs}
\title{Make gRNAs from GeCKO Library}
\usage{
make_gRNAs(seqs, gene = NULL, add2 = list(seq1_5p = "CACCG", seq1_3p =
  "", seq2_5p = "AAAC", seq2_3p = "C"))
}
\arguments{
\item{gene}{}
}
\description{
Make gRNAs from GeCKO Library
}
