% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_gRNAsC.R
\name{make_gRNAsC}
\alias{make_gRNAsC}
\title{Make gRNAs from GeCKO Library}
\usage{
make_gRNAsC(seqs, gene = NULL, add2 = list(seq1_5p = "CACCG", seq1_3p =
  "", seq2_5p = "AAAC", seq2_3p = "C"))
}
\arguments{
\item{add2}{}
}
\description{
Make gRNAs from GeCKO Library
}
