% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{make_sample_score_histograms}
\alias{make_sample_score_histograms}
\title{Histograms of sample scores by class (makes 3 plots)}
\usage{
make_sample_score_histograms(
  scores.df,
  donor0 = "MARS0022.F.16s",
  donor1 = "MARS0043.F.16s",
  title0 = "Healthy (59) Sample Scores",
  title1 = "Asthmatic (36) Sample Scores",
  class1.color = "#DE4968FF",
  class0.color = "#51127CFF",
  both.color = "lightblue",
  class.to.plot,
  x.label = "Sample Score",
  numbins = 10
)
}
\arguments{
\item{scores.df}{dataframe with column named "score" containing numeric score values and "class" column containing class names per sample by row}

\item{donor0}{control class donor sample name (e.g. healthyPatient32)}

\item{donor1}{case class donor sample name (e.g. asthmaPatient19)}

\item{title0}{title for control plot}

\item{title1}{title for case plot}

\item{class1.color}{hex value, or R color name}

\item{class0.color}{hex value, or R color name}

\item{both.color}{hex value, or R color name}

\item{x.label}{x-axis label}

\item{numbins}{breaks for hist}
}
\value{
3 plots inline
}
\description{
Histograms of sample scores by class (makes 3 plots)
}
