% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform.R
\name{perform_NBC}
\alias{perform_NBC}
\title{Use trained model to predict on data}
\usage{
perform_NBC(model, asv_table, sample_data, prob_class1)
}
\arguments{
\item{model}{(object) direct output from train_NBC}

\item{asv_table}{(double) rows= ASV names, cols=sample names}

\item{sample_data}{(character matrix) column 1 = sample names; column 2 = class labels (e.g. "asthmatic" and "healthy")}

\item{prob_class1}{(integer) number samples of test class 1 (class 1 is alphabetically first, class 2 is next) / total number samples see "sample_classes" from train_NBC to check order}
}
\description{
Use trained model to predict on data
}
