% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{PlotAllROCByFeature}
\alias{PlotAllROCByFeature}
\title{Plot a feature ROC Curve}
\usage{
PlotAllROCByFeature(
  sample_data,
  perform_NBC_output,
  train_NBC_output,
  featureAUCsRanked,
  idx1,
  idx2,
  title,
  want_label = TRUE
)
}
\arguments{
\item{sample_data}{character table col1=sample names, col2=class names}

\item{perform_NBC_output}{perform_NBC output}

\item{train_NBC_output}{train_NBC output}

\item{featureAUCsRanked}{vector of AUCs in rank order}

\item{idx1}{integer rank of first feature to plot}

\item{idx2}{integer rank of last feature to plot}

\item{title}{plot title}

\item{want_label}{Boolean value - TRUE will print feature name and AUC on the plot (can only fit up to ~6)}
}
\description{
Plot a feature ROC Curve
}
