% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.R
\name{prior}
\alias{prior}
\title{Gaussian prior function for MAP estimation}
\usage{
prior(alpha, mu, sig)
}
\arguments{
\item{alpha}{beta function parameter}

\item{mu}{mu}

\item{sig}{sigma}
}
\description{
Gaussian prior function for MAP estimation
}
