% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\name{spectre.template}
\alias{spectre.template}
\title{Download the SPECTRE template.}
\usage{
spectre.template()
}
\description{
Download the raster template for SPECTRE layers to your gecko work directory. 
If you have not yet setup a work directory, it will be be setup as if running 
\code{\link[gecko:gecko.setDir]{gecko::gecko.setDir()}} with \code{gisPath = NULL}.
This is a large dataset that is prone to fail by timeout if downloaded 
through R. Instead of using this function you can run gecko.setDir() (if you 
haven't yet) and download the file at 
https://github.com/VascoBranco/spectre.content/raw/main/spectre.template.zip.
Unzip its contents to a folder "./spectretemplate" inside the folder returned by gecko.getDir().
}
\details{
Reads a txt file pointing to where the world GIS files are stored.
}
\examples{
\dontrun{
spectre.template()
}
}
