% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_fn.R
\name{stats}
\alias{stats}
\title{Get a short summary of a given raster segment.}
\usage{
stats(layer, plot = FALSE)
}
\arguments{
\item{layer}{SpatRaster. Raster object, as defined by package terra, with a single layer.}

\item{plot}{logical. If TRUE, a histogram of raster values is drawn.}
}
\value{
data.frame. If plot is TRUE, also outputs a histogram of the layer.
}
\description{
Return a set of descriptive statistics of the given layer,
either a specific one (minimum, q1, median, q3, maximum,
median absolute deviation (mad), mean, standard deviation (sd)) or all of them.
}
\examples{
region = gecko.data("layers")
stats(region[[1]])
}
