% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_nbc.R
\name{train_NBC}
\alias{train_NBC}
\title{NBC Training Function"asv_table" = double\preformatted{    rows= ASV names, cols=sample names
}

"sample_data" = character matrix
column 1 = sample names; column 2 = class labels (e.g. "asthmatic" and "healthy")
"minimum_detection" = integer
is how many samples in which the ASV has to be present in to be included in the training
"min_rel_abund" = integer
sequencing detection limit - the relative abundance value at which we no longer trust the ASV is truly present}
\usage{
train_NBC(asv_table, sample_data, minimum_detection, min_rel_abund)
}
\description{
NBC Training Function"asv_table" = double\preformatted{    rows= ASV names, cols=sample names
}

"sample_data" = character matrix
column 1 = sample names; column 2 = class labels (e.g. "asthmatic" and "healthy")
"minimum_detection" = integer
is how many samples in which the ASV has to be present in to be included in the training
"min_rel_abund" = integer
sequencing detection limit - the relative abundance value at which we no longer trust the ASV is truly present
}
