% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_fn.R
\name{clean}
\alias{clean}
\title{Uniformize raster layers.}
\usage{
clean(layers)
}
\arguments{
\item{layers}{SpatRaster. As defined in package terra, see \code{\link[terra:rast]{terra::rast()}}.}
}
\value{
SpatRaster. Same class as layers.
}
\description{
Crop raster layers to minimum size possible and uniformize \code{NA} values across layers.
}
\details{
Excludes all marginal rows and columns with only \code{NA} values and change values to \code{NA} if they are \code{NA} in any of the layers.
}
\examples{
region = gecko.data("layers")
terra::plot(clean(region))
}
