% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_fn.R
\name{create.lat}
\alias{create.lat}
\title{Create latitude layer.}
\usage{
create.lat(layers)
}
\arguments{
\item{layers}{Raster* object as defined by package raster.}
}
\value{
A RasterLayer object.
}
\description{
Create a layer depicting latitude based on any other.
}
\details{
Using latitude (and longitude) in models may help limiting the extrapolation of the predicted area much beyond known areas.
}
\examples{
data(gecko.layers)
raster::plot(create.lat(gecko.layers[[1]]))
}
