% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_fn.R
\name{create.north}
\alias{create.north}
\title{Create northness layer.}
\usage{
create.north(layers)
}
\arguments{
\item{layers}{SpatRaster. A layer of elevation (a digital elevation model - DEM). 
As defined in package terra, see \code{\link[terra:rast]{terra::rast()}}.}
}
\value{
SpatRaster.
}
\description{
Create a layer depicting northness based on an elevation layer.
}
\details{
Using elevation, aspect can be calculated. Yet, it is a circular variable (0 = 360) and has to be converted to northness and eastness to be useful for modelling.
}
\examples{
region = gecko.data("layers")
terra::plot(create.north(region[[3]]))
}
